Adds unknown as a basic type in place of any and cautions against using any.
`any` should really only be used as a last resort or when porting JS code.  The docs prior to this PR held `any` in a pretty positive light and suggested some uses for it that really weren't appropriate (`unknown` would have been better).  This PR swaps in `unknown` in the spot `any` previously held in the basic types section, moves `any` to the bottom of the list, and includes language that discourages the usage of `any` in favor of other options like `unknown`, Mapped Types, and Index Types.  Fixed up some references to `any` as well, redirecting them to `unknown`.

Also adjusted the wording on one of the `never` comments while I was in here.