Fixed example that should fail
The text tells us:
`The above workaround will work as long as you have a common property between squareOptions and SquareConfig. In this example, it was the property width. It will however, fail if the variable does not have any common object property. For example:`

and then comes an example that in fact not fail, but works fine. Changed the example to fail instead, as intended in its description text.