Add Error comment for accessing method of subclass
Currently an error occurs at the line:
```
department.generateReports();
```

This happens because department is not of type `AccountingDepartment` hence can't access the `generateReports` method. This is however harder to understand because:

1. The type is declared in the start a few lines above to be Department and later object is assigned from constructor of `AccountingDepartment`.
2. There is no comment to point out that the error will occur and neither is it described in the text.

This confused me for a while to understand that the type didn't change and to get the method to work, department should be initialized with the correct type: `AccountingDepartment`. Even though the error message displayed in the code is very similar, adding the comment makes it known to the reader that this error is expected as this is not discussed in the text.

This PR adds the comment necessary to note that the error is expected and explains why the error occurred.

 Hopefully this will cause less confusion for someone else.