Upper case file name fails
If link is clicked and is loaded - I am assuming - by Javascript, the upper case makes the page fail to load. Changing the leading character to lower-case makes it work. Opening the link in a new window also works, as that uses HTML fetch which is not case-sensitive.

There is the same problem in the search index (when searching for any Symbol's methods), but I don't know how to fix that.