Typo in input and output
While I was scrolling trought TS docs i found this realy diferent thing that is the table with a <div> and <div> output for react native. Its been a while since I've not used React Native but I remeber that we used <View> that could transpile to  UIView, <div>, android.view depending on your platform, but you should refer to this block of code as <View> like discussed in this doc (RN View Doc) [https://reactnative.dev/docs/view].
If i'm wrong on this, I would like to say sorry.  This is my first Pull request to making TypeScript better. If there is any things that i could do to make this pr better I gladly would change :) thx and good holidays