Clarify property renaming example
This section steps the reader through how properties can 
be renamed. Here, the point is that types must be added 
separately, after the entire destructuring-with-rename has
been accomplished. But as currently written, this line has 
dropped the property renaming. The example is clearer if 
the property renaming remains, so the only difference is
the introduction of the type annotations.