CommonJs file module's path should be relative
If it is not relative it is considered a core module instead of a file module. Here 'maths.js' is a file module, note a core module that we are requiring from `node_modules` or from the standard library.  For more information please refer the file module doc: https://nodejs.org/api/modules.html#file-modules