Update Iterators and Generators.md
I suggested that the removed section in the Pull Request be merged because it is showing an incorrect usage of the `Set` object according to [MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set).

In theory this what the author wanted to communicate with this section makes sense (if they were teaching JavaScript), if you ran that example in JavaScript land it would work and make sense. But with TypeScript it will throw an error and not even compile depending on how 'strict' the compiler settings are . Therefore it causes only confusion and should be removed in my opinion.

Perhaps there's a better way to to get the same point across that the author wanted to, but I couldn't think of any so far. I could also try to rewrite a similar / different suggestion if anyone had an idea for that. 

There's an open issue from 2020 which shows the same confusion I had when I first saw this example [link](https://github.com/microsoft/TypeScript-Website/issues/909)

The error thrown by TS for this example: 
```
Element implicitly has an 'any' type because expression of type '"species"' can't be used to index type 'Set<string>'.
  Property 'species' does not exist on type 'Set<string>'.
```