Update Generics.md | Bees have six legs, not four.
I updated the last example given to use Lions and Elephants instead of Bees and Lions. I did this because the previous example sets Bees to have four legs. Bees are insects, and thus have six legs. A pedantic point, but I think it's helpful when examples are accurate to real-world scenarios.