Note CLI ignores tsconfig if files are specified
Added verbatim copy of the sentence from [here](https://github.com/microsoft/TypeScript-Website/blob/v2/packages/documentation/copy/en/project-config/tsconfig.json.md?plain=1#L23).

Redundancy in the documentation is probably not the best thing, however this is such an important note about the CLI behaviour. It doesn't seem to be reasonable to look for information about CLI behaviour on another page. If it can only be added in one place, the CLI page would be the better candidate, but coming from the perspective of the readers' expectations, it is reasonable to have this sentence on both the tsconfig and the cli pages.

 Also current first sentence suggests that providing files and using tsconfig is possible at the same time. While both parts of that sentence is true, they are not true at the same time, so an or between them helps to avoid wrong expectations.