copy/en/get-started: make `type/value` distinction
This change tries to make the distinction between a type and
a value more explicit. Indeed, in the TS world, something such as
`"a string"` could be well regarded as a type, but it is more naturally
known as a value. Given that this piece of information is in a learners'
section, I believe that the change is worth making as it could be
confusing.

Signed-off-by: Emmanuel Chee-zaram Okeke <cheezaram.okeke@gmail.com>