Explicitly call out which line is "a template literal type"
I'm not confident in this change, but I'm a little worried that the `type PopularTopLevelDomains` could be misconstrued as the template literal type in the example since it's the first line. 

I tried to circumvent this by swapping the order of those first two lines, but I didn't like how it used `PopularTopLevelDomains` before it was defined. Therefore, this seemed like the lesser of two evils.