Updated twoslasher custom tag matcher to allow whitespace
Currently, there is a bug in the custom tag matcher which doesn't show up in the commented regex example due to a subtle difference between the example regex and the production code regex. The linked regex sample ( for convenience: https://regex101.com/r/8B2Wwh/1 ) differs from the production code in a subtle way: it omitted the starting `^` used in the production regex. This subtle difference means that in the linked example, comments indented with spaces or tabs are still matched. However, the production regex begins with `/^\/\/`, which will not match comments that have been indented as it requires `//` be the first characters of the line. Here's an example regex showing that an indented commented fails to match: https://regex101.com/r/fkqmfp/1

This is an issue however because it is valid *and common* for a comment to be intended to line up with the code being commented. In fact, this is a standard behavior of most automated formatting tools.

With this change, custom tag comments that are indented will be matched by the `valuedConfigRegexp`: see https://regex101.com/r/0pHWps/1